Update draft-fox-lsr-ospf-terminology.xml
Updated my address, fixed one incorrect reference, and added a new name for the MS-bit in the database exchange packet. The bit name change may seem obvious but since it also had an "official" abbreviation of MS, I wanted to give it a new one, L. I used L instead of LF because I think MS was chosen in the first place not necessarilyi to mean master/slave but to distinguish it from the More (M) bit.